+++ src/datasources/data-frame/DataFrameDataSource.ts
@@ -30,7 +30,14 @@ export class DataFrameDataSource extends DataFrameDataSourceBase {
   ) {
     super(instanceSettings, backendSrv, templateSrv);
 
-    this.datasource = new DataFrameDataSourceV2(instanceSettings, backendSrv, templateSrv);
+    const featureToggles = instanceSettings.jsonData?.featureToggles;
+
+    this.datasource = new DataFrameDataSourceV2(
+        instanceSettings,
+        backendSrv,
+        templateSrv,
+        featureToggles
+      );
   }
 
   public get defaultQuery(): Required<Omit<DataFrameQuery, keyof DataQuery>> {
@@ -104,8 +111,11 @@ export class DataFrameDataSource extends DataFrameDataSourceBase {
     return this.datasource.queryTables(query, take, projection);
   }
 
-  public processQuery(query: DataFrameDataQuery): ValidDataFrameQuery {
-    return this.datasource.processQuery(query);
+  public processQuery(
+    query: DataFrameDataQuery,
+    queries: DataFrameDataQuery[] = []
+  ): ValidDataFrameQuery {
+    return this.datasource.processQuery(query, queries);
   }
 
   public processVariableQuery(query: DataFrameVariableQuery): ValidDataFrameVariableQuery {